refactor: rename wikidata mappers folder to utils and update import paths after removing mapper files.
