--- backend/routes/layersRoutes.js
@@ -1,9 +1,0 @@
-import express from "express";
-import { listLayers, listLayerEvents } from "./layers.controller.js";
-
-const router = express.Router();
-
-router.get("/", listLayers);
-router.get("/:id/events", listLayerEvents);
-
-export default router;
